% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{circleplot}
\alias{circleplot}
\title{Preparation for circle plot}
\usage{
circleplot(preplot, scoregene)
}
\arguments{
\item{preplot}{a list as obtained from the function plotcrosstalk}

\item{scoregene}{a score for each gene with values included between -10 e +10}
}
\value{
a list with correlation matrix and gene set for each gene
}
\description{
circleplot function takes as input data derived by the function plotcrosstalk and plOt a circle plot.
}
\examples{
formatplot<-plotcrosstalk(pathway_plot=pathway[1:6],gs_expre=tumo)
score<-runif(length(formatplot[[2]]), min=-10, max=+10)
circleplot(preplot=formatplot,scoregene=score)
}
