% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{dsscorecrtlk}
\alias{dsscorecrtlk}
\title{For TCGA data get human pathway data and creates a measure of discriminating score among pathways}
\usage{
dsscorecrtlk(dataFilt, pathway_exp)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway_exp}{a list of pathway data}
}
\value{
a matrix value for each pathway
}
\description{
dsscorecrtlk creates a matrix with  discriminating score for pathways
}
\examples{
cross_talk_st_dv<-dsscorecrtlk(dataFilt=tumo[,1:2],pathway_exp=pathway[1:5])
}
