% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{eucdistcrtlk}
\alias{eucdistcrtlk}
\title{For TCGA data get human pathway data and creates a measure of cross-talk among pathways}
\usage{
eucdistcrtlk(dataFilt, pathway_exp)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway_exp}{list of pathway data}
}
\value{
a matrix value for each pathway
}
\description{
eucdistcrtlk creates a matrix with euclidean distance for pairwise pathways
}
\examples{
score_euc_dista_t<-eucdistcrtlk(dataFilt=tumo[,1:2],pathway_exp=pathway[1:5])
}
