% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{GE_matrix}
\alias{GE_matrix}
\title{Get human KEGG pathway data and a gene expression matrix in order to obtain a list with the gene expression for only pathways given in input .}
\usage{
GE_matrix(DataMatrix, genes.by.pathway)
}
\arguments{
\item{DataMatrix}{gene expression matrix (eg.TCGA data)}

\item{genes.by.pathway}{a list of pathway data as provided by GetData and ConvertedID_genes}
}
\value{
a list for each pathway ( gene expression level belong to that pathway)
}
\description{
GE_matrix creates a list of gene expression for pathways given by the user.
}
\examples{
list_path_gene<-GE_matrix(DataMatrix=tumo[,1:2],genes.by.pathway=pathway[1:5])
}
