% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{GE_matrix_mean}
\alias{GE_matrix_mean}
\title{Get human KEGG pathway data and a gene expression matrix in order to obtain a matrix with the mean gene expression for only pathways given in input .}
\usage{
GE_matrix_mean(DataMatrix, genes.by.pathway)
}
\arguments{
\item{DataMatrix}{gene expression matrix (eg.TCGA data)}

\item{genes.by.pathway}{list of pathway data as provided by getKEGGdata}
}
\value{
a matrix for each pathway (mean gene expression level belong to that pathway)
}
\description{
GE_matrix creates a matrix of mean gene expression levels for pathways given by the user.
}
\examples{
list_path_plot<-GE_matrix_mean(DataMatrix=tumo[,1:2],genes.by.pathway=pathway[1:5])
}
