% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{GetData}
\alias{GetData}
\title{Get general information inside pathways.}
\usage{
GetData(species, pathwaydb)
}
\arguments{
\item{species}{variable. The user can select the species of interest from SELECT_path_species(path_spec)}

\item{pathwaydb}{variable. The user can select the pathway database of interest from SELECT_path_graphite(path_spec)}
}
\value{
a list of pathways
}
\description{
GetData creates a list with genes inside the pathways.
}
\examples{
\dontrun{
species="hsapiens"
pathwaydb="pharmgkb"
path<-GetData(species,pathwaydb)}
}
