% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{GetPathData}
\alias{GetPathData}
\title{Get  genes inside pathways.}
\usage{
GetPathData(path_ALL)
}
\arguments{
\item{path_ALL}{variable. The user can select the variable as obtained by  GetData function}
}
\value{
a list of pathways
}
\description{
GetPathData creates a list of genes inside the pathways.
}
\examples{
pathway_ALL_GENE<-GetPathData(path_ALL=path[1:3])
}
