% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{GetPathNet}
\alias{GetPathNet}
\title{Get  interacting genes inside pathways.}
\usage{
GetPathNet(path_ALL)
}
\arguments{
\item{path_ALL}{variable. The user can select the variable as obtained by  GetData function}
}
\value{
a list of pathways
}
\description{
GetPathNet creates a list of genes inside the pathways.
}
\examples{
pathway_net<-GetPathNet(path_ALL=path[1:3])
}
