% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{matrix_plot}
\alias{matrix_plot}
\title{Get human KEGG pathway data and a gene expression matrix in order to obtain a matrix with the mean gene expression for only pathways given in input .}
\usage{
matrix_plot(DataMatrix, pathway)
}
\arguments{
\item{DataMatrix}{gene expression matrix (eg.TCGA data)}

\item{pathway}{pathway data as provided by getKEGGdata}
}
\value{
a matrix for each pathway (mean gene expression level belong to that pathway)
}
\description{
GE_matrix creates a matrix of mean gene expression for pathways given by the user.
}
\examples{
list_path_plot<-matrix_plot(DataMatrix=tumo[,1:2],pathway=path)
}
