% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{pathnet}
\alias{pathnet}
\title{Get human KEGG pathway data and creates a network data.}
\usage{
pathnet(genes.by.pathway, data)
}
\arguments{
\item{genes.by.pathway}{a list of pathway data as provided by ConvertedIDgenes}

\item{data}{a list of network data as provided by getNETdata}
}
\value{
a list of network data for each pathway (interacting genes belong to that pathway)
}
\description{
pathnet creates a list of network data for each human pathway. The network data will be generated when interacting genes belong to that pathway.
}
\examples{
lista_net<-pathnet(genes.by.pathway=pathway[1:5],data=netw)
}
