% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{plotcrosstalk}
\alias{plotcrosstalk}
\title{Preparation for plotting cross-talk}
\usage{
plotcrosstalk(pathway_plot, gs_expre)
}
\arguments{
\item{pathway_plot}{pathway}

\item{gs_expre}{a gene expression matrix}
}
\value{
a list with correlation matrix and gene set for each gene
}
\description{
plot_cross_talk function takes as input pathway data and prepares the data to visualize (e.g. ggplot2, qqgraph, igraph)
}
\examples{
formatplot<-plotcrosstalk(pathway_plot=pathway[1:6],gs_expre=tumo)
}
