% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{SELECTpathspecies}
\alias{SELECTpathspecies}
\title{List of species}
\usage{
SELECTpathspecies(path_spec)
}
\arguments{
\item{path_spec}{variable}
}
\description{
List of species for network
}
\examples{
\dontrun{m<-SELECTpathspecies(path_spec)
}
}
