% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarBioTrek.r
\docType{package}
\name{StarBioTrek}
\alias{StarBioTrek}
\alias{StarBioTrek-package}
\title{Download data}
\description{
StarBioTrek allows you to Download data of samples from StarBioTrek
}
\details{
The functions you're likely to need from \pkg{StarBioTrek} is
\code{path_star}
Otherwise refer to the vignettes to see
how to format the documentation.
}
