% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{stdv}
\alias{stdv}
\title{For TCGA data get human pathway data and creates a measure of standard deviations among pathways}
\usage{
stdv(gslist)
}
\arguments{
\item{gslist}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
stdv creates a matrix with standard deviation for pathways
}
\examples{
list_path_gene<-GE_matrix(DataMatrix=tumo[,1:2],genes.by.pathway=pathway[1:5])
score_stdev<-stdv(gslist=list_path_gene)
}
